# backend/openapi.yaml
openapi: 3.1.0
...
info:
  title: LoopBuy Backend API
  version: 1.0.0
  description: >-
    REST API for LoopBuy accounts, marketplace listings, social features,
    scam screening, recommendations, and the RAG shopping assistant.
servers:
  - url: http://localhost:8090
    description: Docker Compose local API
security: []
...
paths:
...
  /api/v1/categories/{identifier}:
    parameters:
      - $ref: '#/components/parameters/CategoryIdentifier'
    get:
      tags: [Categories]
      operationId: getCategory
      summary: Get an active category by ID or slug
      responses:
        '200':
          description: Category details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Categories]
      operationId: updateCategory
      summary: Update a category
      description: Administrator role is required and `identifier` must be numeric.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryPatchRequest'
      responses:
        '200':
          description: Updated category.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Categories]
      operationId: deleteCategory
      summary: Deactivate a category
      description: Administrator role is required and `identifier` must be numeric.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Category deactivated.
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/listings:
    get:
      tags: [Listings]
      operationId: listListings
      summary: Browse listings
      security:
        - bearerAuth: []
        - {}
      parameters:
        - name: q
          in: query
          description: MySQL natural-language full-text search. `product_search` is accepted as a legacy alias.
          schema:
            type: string
            maxLength: 120
        - name: product_search
          in: query
          description: Legacy alias used only when `q` is empty.
          schema:
            type: string
            maxLength: 120
        - name: category
          in: query
          description: Category ID or slug.
          schema:
            type: string
            maxLength: 100
        - name: min_price
          in: query
          schema:
            type: number
            minimum: 0
        - name: max_price
          in: query
          schema:
            type: number
            minimum: 0
        - name: condition
          in: query
          schema:
            $ref: '#/components/schemas/ItemCondition'
        - name: location
          in: query
          schema:
            type: string
            maxLength: 120
        - name: seller_id
          in: query
          schema:
            type: integer
            format: int64
            minimum: 1
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, price-low, price-high]
            default: newest
        - $ref: '#/components/parameters/ListingsLimit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Matching listings. Anonymous callers see only active/approved items.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingPage'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Listings]
      operationId: createListing
      summary: Create and screen a listing
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingCreateRequest'
      responses:
        '201':
          description: Listing created with its current moderation state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/listings/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Listings]
      operationId: getListing
      summary: Get a listing
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Listing details. Non-public states are visible only to the seller or an admin.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      tags: [Listings]
      operationId: updateListing
      summary: Update and rescreen a listing
      description: >-
        The JSON body must include the `revision` from the latest listing
        representation. Missing or zero revision returns 428. If another
        mutation has advanced the listing revision, the stale update returns
        409 and the client must reload before retrying.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingUpdateRequest'
      responses:
        '200':
          description: Updated listing and moderation state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        '409':
          description: The listing changed while scam screening was running, so the stale update was rejected.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '428':
          description: The request body omitted a positive current listing revision.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Listings]
      operationId: deleteListing
      summary: Archive a listing
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Listing status changed to archived and vector deletion queued.
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/listings/{id}/status:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Listings]
      operationId: updateListingStatus
      summary: Change listing status
      description: >-
        Activation requires moderation status approved. A seller cannot reactivate
        a sold listing; an administrator can.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingStatusRequest'
      responses:
        '200':
          description: Listing with the new status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listing'
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/listings/{id}/images:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Listings]
      operationId: listListingImages
      summary: List images for a visible listing
      description: >-
        Non-public listing images are visible only to the seller, an admin, or
        a moderator.
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Listing images in primary/sort order.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingImageList'
        default:
          $ref: '#/components/responses/Problem'
    post:
      tags: [Listings]
      operationId: addListingImage
      summary: Import an external listing image (administrator only)
      description: >-
        Legacy/import endpoint restricted to administrators. Sellers use the
        multipart local-storage endpoint so marketplace pages never depend on
        third-party image hosts or leak visitor IP addresses to them.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImageInput'
      responses:
        '201':
          description: Image created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingImage'
        default:
          $ref: '#/components/responses/Problem'
...
  /api/v1/listings/{id}/images/{imageId}:
    parameters:
      - $ref: '#/components/parameters/Id'
      - $ref: '#/components/parameters/ImageId'
    patch:
      tags: [Listings]
      operationId: updateListingImage
      summary: Update listing image metadata
      description: >-
        Sellers may change sort_order and is_primary but cannot replace
        image_url. Administrators may also provide an HTTP(S) URL for a legacy
        import. To replace seller media, upload a new local image and delete the
        old object.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingImageUpdate'
      responses:
        '200':
          description: Updated image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingImage'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Listings]
      operationId: deleteListingImage
      summary: Delete a listing image
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Image deleted.
        default:
          $ref: '#/components/responses/Problem'
  /api/v1/listings/{id}/scam-assessments:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Listings]
      operationId: assessListing
      summary: Rerun scam screening
      security:
        - bearerAuth: []
      responses:
        '201':
          description: New assessment and updated listing moderation state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScamAssessment'
        '409':
          description: The listing changed while reassessment was running, so the stale moderation result was rejected.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: HS256 access token issued by register, login, or refresh.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    CategoryIdentifier:
      name: identifier
      in: path
      required: true
      description: Numeric category ID or slug for GET; numeric ID for PATCH/DELETE.
      schema:
        type: string
        pattern: '^(?:[1-9][0-9]*|[a-z0-9]+(?:-[a-z0-9]+)*)$'
...
    ListingsLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 10000
        default: 0
  responses:
    Problem:
      description: Error response. Typical statuses are 400, 401, 403, 404, 409, 422, 428, 500, or 503.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CategoryPatchRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 100
        slug:
          type: string
          maxLength: 100
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        is_active:
          type: boolean
...
    ItemCondition:
      type: string
      enum: [new, like_new, good, fair]
...
    Category:
      type: object
      additionalProperties: false
      required: [category_id, name, slug, is_active, created_at]
      properties:
        category_id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
...